Autonomy Software/requirements.txt
ArducamDepthCamera==0.1.21
bidict==0.23.1
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
eventlet==0.40.4
fastcrc==0.3.4
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
greenlet==3.3.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==1.26.4
opencv-python==4.11.0.86
pymavlink==2.4.49
pyserial==3.5
python-engineio==4.12.3
python-socketio==5.15.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.3.2
